+++ lib/widgets/home/statistics/overall.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: must_be_immutable, prefer_final_fields, unused_field, prefer_const_constructors, avoid_print
 
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:learning_english_app/models/practice/result.dart';
 import 'package:learning_english_app/resources/firebase_handle.dart';
 import 'package:learning_english_app/screens/statistic/history_screen.dart';
@@ -83,12 +82,11 @@ class Overall extends StatelessWidget {
             width: double.infinity,
             child: ElevatedButton(
               onPressed: () {
-                // showModalBottomSheet(
-                //   context: context,
-                //   builder: (context) => HistoryScreen(),
-                //   isScrollControlled: true,
-                // );
-                Get.to(HistoryScreen());
+                showModalBottomSheet(
+                  context: context,
+                  builder: (context) => HistoryScreen(),
+                  isScrollControlled: true,
+                );
               },
               child: Text("View history"),
               style: ButtonStyle(
